Separate out almost all code in tests/mod
Move this code into 3 new submodules: helpers, normal_tests,
setup_teardown_tests
